@@ -2,9 +2,6 @@
 
 namespace AnyContent\Client;
 
-use CMDL\Util;
-use AnyContent\Client\File;
-
 class Folder
 {
 
@@ -13,13 +10,13 @@ class Folder
     protected $subFolders = array();
 
 
-    public function __construct($path, $result)
+    public function __construct($path, $data)
     {
 
         $path = trim($path, '/');
 
         $this->path = $path;
-        foreach ($result['files'] as $file)
+        foreach ($data['files'] as $file)
         {
             $this->files[$file['id']] = new File($this, $file['id'], $file['name'], $file['type'], $file['urls'], $file['size'], $file['timestamp_lastchange']);
             if ($file['type'] == 'image' AND array_key_exists('width', $file) AND array_key_exists('height', $file))
@@ -28,9 +25,13 @@ class Folder
                 $this->files[$file['id']]->setWidth($file['width']);
                 $this->files[$file['id']]->setHeight($file['height']);
             }
+            if (array_key_exists('url', $file))
+            {
+                $this->files[$file['id']]->addUrl('default', $file['url']);
+            }
         }
 
-        foreach ($result['folders'] as $folder)
+        foreach ($data['folder'] as $folder)
         {
             $this->subFolders[$this->path . '/' . $folder] = $folder;
         }
