@@ -2,47 +2,69 @@
 
 namespace AnyContent\Client;
 
-use CMDL\Parser;
-
-use AnyContent\Client\Client;
-use AnyContent\Client\UserInfo;
-use AnyContent\Client\Record;
+use AnyContent\Connection\Configuration\ContentArchiveConfiguration;
+use AnyContent\Connection\ContentArchiveReadWriteConnection;
+use KVMLogger\KVMLoggerFactory;
+use KVMLogger\KVMLogger;
+use Symfony\Component\Filesystem\Filesystem;
 
 class UserInfoTest extends \PHPUnit_Framework_TestCase
 {
 
-    /**
-     * @var $client Client
-     */
-    public $client = null;
+    /** @var  ContentArchiveReadWriteConnection */
+    public $connection;
+
+    /** @var  Repository */
+    public $repository;
+
+
+    public static function setUpBeforeClass()
+    {
+        $target = __DIR__ . '/../../../tmp/ExampleContentArchive';
+        $source = __DIR__ . '/../../resources/ContentArchiveExample2';
+
+        $fs = new Filesystem();
+
+        if (file_exists($target))
+        {
+            $fs->remove($target);
+        }
+
+        $fs->mirror($source, $target);
+
+    }
 
 
     public function setUp()
     {
+        $target = __DIR__ . '/../../../tmp/ExampleContentArchive';
 
-        global $testWithCaching;
+        $configuration = new ContentArchiveConfiguration();
 
-        $cache = null;
-        if ($testWithCaching)
-        {
-            $cache = new \Doctrine\Common\Cache\ApcCache();
-        }
+        $configuration->setContentArchiveFolder($target);
 
-        // Connect to repository
-        $client = new Client('http://acrs.github.dev/1/example', null, null, 'Basic', $cache);
-        $client->setUserInfo(new UserInfo('john.doe@example.org', 'John', 'Doe'));
-        $this->client = $client;
+        $connection = $configuration->createReadWriteConnection();
+
+        $this->connection = $connection;
+
+        $this->repository = new Repository('phpunit',$this->connection);
+
+        KVMLoggerFactory::createWithKLogger(__DIR__.'/../../../tmp');
+
+
     }
 
 
-    public function testCreationInfo()
+    public function testUserInfo()
     {
-        $cmdl = $this->client->getCMDL('example01');
+        $this->repository->setUserInfo(new UserInfo('john.doe@example.org', 'John', 'Doe'));
 
-        $contentTypeDefinition = Parser::parseCMDLString($cmdl);
-        $contentTypeDefinition->setName('example01');
+        $this->repository->selectContentType('example01');
 
-        $record = $this->client->getRecord($contentTypeDefinition, 1);
+        $record = $this->repository->createRecord('New Record ');
+
+        $id = $this->repository->saveRecord($record);
+        $this->assertEquals(1, $id);
 
         $this->assertInstanceOf('AnyContent\Client\UserInfo', $record->getCreationUserInfo());
         $this->assertInstanceOf('AnyContent\Client\UserInfo', $record->getLastChangeUserInfo());
@@ -54,5 +76,55 @@ class UserInfoTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('John', $userinfo->getFirstname());
         $this->assertEquals('Doe', $userinfo->getLastname());
         $this->assertTrue($userinfo->userNameIsAnEmailAddress());
+
+    }
+
+
+    public function testUserNewConnectionInfo()
+    {
+
+        $this->repository->selectContentType('example01');
+
+        $record = $this->repository->getRecord(1);
+
+        $this->assertInstanceOf('AnyContent\Client\UserInfo', $record->getCreationUserInfo());
+        $this->assertInstanceOf('AnyContent\Client\UserInfo', $record->getLastChangeUserInfo());
+
+        /** @var UserInfo $userinfo */
+        $userinfo = $record->getCreationUserInfo();
+
+        $this->assertEquals('john.doe@example.org', $userinfo->getUsername());
+        $this->assertEquals('John', $userinfo->getFirstname());
+        $this->assertEquals('Doe', $userinfo->getLastname());
+        $this->assertTrue($userinfo->userNameIsAnEmailAddress());
+    }
+
+
+    public function testChangeUserInfo()
+    {
+        $this->repository->setUserInfo(new UserInfo('john.tester@example.org', 'John', 'Tester'));
+
+        $this->repository->selectContentType('example01');
+
+        $record = $this->repository->getRecord(1);
+
+        $this->repository->saveRecord($record);
+
+        /** @var UserInfo $userinfo */
+        $userinfo = $record->getCreationUserInfo();
+
+        $this->assertEquals('john.doe@example.org', $userinfo->getUsername());
+        $this->assertEquals('John', $userinfo->getFirstname());
+        $this->assertEquals('Doe', $userinfo->getLastname());
+        $this->assertTrue($userinfo->userNameIsAnEmailAddress());
+
+        /** @var UserInfo $userinfo */
+        $userinfo = $record->getLastChangeUserInfo();
+
+        $this->assertEquals('john.tester@example.org', $userinfo->getUsername());
+        $this->assertEquals('John', $userinfo->getFirstname());
+        $this->assertEquals('Tester', $userinfo->getLastname());
+        $this->assertTrue($userinfo->userNameIsAnEmailAddress());
+
     }
 }
